requirements.txt
asgiref==3.5.2
Django==4.0.4
django-crispy-forms==1.14.0
Pillow==9.1.1
sqlparse==0.4.2
